x Cosmetic/warnings removal update
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@4708 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/Multiple/File_Ancillary.cpp
@@ -322,7 +322,7 @@ void File_Ancillary::Header_Parse()
 
     //Test (in some container formats, Cheksum is present sometimes)
     bool WithChecksum_Temp=WithChecksum;
-    if (!MustSynchronize && !WithChecksum && (3+DataCount+1)*(WithTenBit?2:1)==Buffer_Size)
+    if (!MustSynchronize && !WithChecksum && (size_t)((3+DataCount+1)*(WithTenBit?2:1))==Buffer_Size)
         WithChecksum_Temp=true; 
     
     //Filling
